added-form
t this juncture of development I implemented the following:

Create a method called current_user in the ApplicationController to make current
user data available to all controllers.
It returns the first user from the database.
Create forms to perform the following functions:
Creates a Post on behalf of the current_user
Create a comment on behalf of the current_user
Allow Users to add likes to Posts.